+++ shaders/style_vanilla.glsl
@@ -323,7 +323,7 @@
 
 #if END_SKYLIGHT_RED == -1
 #undef END_SKYLIGHT_RED
-										#define END_SKYLIGHT_RED 1.05
+										#define END_SKYLIGHT_RED 1.1
 #endif
 #if END_SKYLIGHT_GREEN == -1
 #undef END_SKYLIGHT_GREEN
@@ -331,7 +331,7 @@
 #endif
 #if END_SKYLIGHT_BLUE == -1
 #undef END_SKYLIGHT_BLUE
-										#define END_SKYLIGHT_BLUE 1.1
+										#define END_SKYLIGHT_BLUE 1.15
 #endif
 #if END_SKYLIGHT_BRIGHTNESS == -1
 #undef END_SKYLIGHT_BRIGHTNESS
@@ -339,7 +339,7 @@
 #endif
 #if END_AMBIENT_RED == -1
 #undef END_AMBIENT_RED
-										#define END_AMBIENT_RED 1.05
+										#define END_AMBIENT_RED 1.1
 #endif
 #if END_AMBIENT_GREEN == -1
 #undef END_AMBIENT_GREEN
@@ -347,11 +347,11 @@
 #endif
 #if END_AMBIENT_BLUE == -1
 #undef END_AMBIENT_BLUE
-										#define END_AMBIENT_BLUE 1.05
+										#define END_AMBIENT_BLUE 1.1
 #endif
 #if END_AMBIENT_BRIGHTNESS == -1
 #undef END_AMBIENT_BRIGHTNESS
-										#define END_AMBIENT_BRIGHTNESS 0.45
+										#define END_AMBIENT_BRIGHTNESS 0.5
 #endif